+++ include/protowire/pxf.h
@@ -27,7 +27,7 @@
 #include "protowire/pxf/parser.h"  // Document, Parse
 #include "protowire/pxf/result.h"
 #include "protowire/pxf/schema.h"        // ValidateDescriptor, Violation
-#include "protowire/pxf/table_reader.h"  // TableReader, BindRow
+#include "protowire/pxf/dataset_reader.h"  // DatasetReader, BindRow
 
 namespace protowire::pxf {